+++ src/App.jsx
@@ -1,27 +1,27 @@
 import "./App.css";
-import AuthLogin1 from "./AuthLogin1";
-import AuthLogin from "./AuthLongin";
-import CattleManagement from "./cattleManagement";
-import CattleQuery from "./CattleQuery";
+import AuthLogin1 from "./pages/AuthLogin1";
+import AuthLogin from "./pages/AuthLongin";
+import CattleManagement from "./pages/CattleManagement";
+import CattleQuery from "./pages/CattleQuery";
 import Into1 from "./pages/Into1";
 import Into2 from "./pages/Into2";
 import Into3 from "./pages/Into3";
-import ReportQuery from "./ReportQuery";
-import SelectLanguage from "./SelectLanguage";
-import SelectYourCrop from "./SelectYourCrop";
-import SoilReport from "./SoilReport";
-import SoilReport3 from "./SoilReport3";
-import SoilReportSubmit from "./SoilReportSubmit";
+import ReportQuery from "./pages/ReportQuery";
+import SelectLanguage from "./pages/SelectLanguage";
+import SelectYourCrop from "./pages/SelectYourCrop";
+import SoilReport from "./pages/SoilReport";
+import SoilReport3 from "./pages/SoilReport3";
+import SoilReportSubmit from "./pages/SoilReportSubmit";
 
 import Splash from "./pages/Splash";
 import { Route, Routes } from "react-router-dom";
-import Step1 from "./Step1";
-import Step2 from "./Step2";
-import Step3 from "./Step3";
-import Step4 from "./Step4";
-import Step5 from "./Step5";
-import Step6 from "./Step6";
-import Step7 from "./Step7";
+import Step1 from "./pages/Step1";
+import Step2 from "./pages/Step2";
+import Step3 from "./pages/Step3";
+import Step4 from "./pages/Step4";
+import Step5 from "./pages/Step5";
+import Step6 from "./pages/Step6";
+import Step7 from "./pages/Step7";
 import { IntroStepper } from "./components/IntroStepper";
 import intro1Bg from './assets/into1.png';
 import WelcomePage from "./pages/WelcomePage";
