+++ src/components/StepPage.jsx
@@ -1,13 +1,13 @@
 // src/components/StepPage.jsx
 
 import React from 'react';
-import savpalImg from "./assets/savpal.png";
+import savpalImg from "../assets/savpal.png";
 
 const StepPage = ({ title, instruction, children }) => {
   return (
-    <div className="h-screen flex flex-col items-center bg-gray-200 p-2">
+    <div className="h-screen flex flex-col mx-auto w-full items-center">
       {/* Main card container */}
-      <div className="w-full flex-grow rounded-xl overflow-hidden shadow-lg bg-gray-100 flex flex-col">
+      <div className="w-full flex-grow shadow-md bg-gray-100 flex flex-col">
         {/* Header with gradient */}
         <div
           className="h-24"
@@ -17,7 +17,7 @@ const StepPage = ({ title, instruction, children }) => {
         ></div>
 
         {/* Logo */}
-        <div className="flex justify-center">
+        <div className="flex max-w-md mx-auto w-full justify-center">
           <div
             className="w-40 h-20 mt-2 bg-no-repeat bg-contain"
             style={{ backgroundImage: `url(${savpalImg})` }}
@@ -25,10 +25,10 @@ const StepPage = ({ title, instruction, children }) => {
         </div>
 
         {/* Scrollable content area */}
-        <div className="p-4 flex overflow-y-auto flex-col gap-4">
+        <div className="p-2 flex  flex-col  max-w-md mx-auto gap-2">
           <h3 className="text-base font-semibold text-black">{title}</h3>
           <span className="text-black text-sm">Instruction:</span>
-          <p className="text-gray-600 text-sm">{instruction}</p>
+          <p className="text-gray-600 text-sm ">{instruction}</p>
           {children}
         </div>
