+++ src/pages/ReportQuery.jsx
@@ -1,13 +1,13 @@
-import yourLogo from "./assets/soil.png";
-import eyebIcon from "./assets/eyeblue.png";
-import deleteIcon from "./assets/delete.png";
-import downloadIcon from "./assets/down.png";
-import calendar from "./assets/calendardate.png";
+import yourLogo from "../assets/soil.png";
+import eyebIcon from "../assets/eyeblue.png";
+import deleteIcon from "../assets/delete.png";
+import downloadIcon from "../assets/down.png";
+import calendar from "../assets/calendardate.png";
 
 
 const ReportQuery = () => {
   return (
-    <div className="h-screen flex justify-center flex-col bg-gray-200 items-center">
+    <div className="h-screen flex justify-center flex-col max-w-md mx-auto bg-gray-200 items-center">
       <div className="w-full h-full shadow-lg overflow-y-auto">
         <div
           className="h-24"
