+++ src/pages/SelectLanguage.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import LinkButton from "./components/LinkButton";
+import LinkButton from "../components/LinkButton";
 
 const SelectLanguage = () => {
   const [selectedLanguage, setSelectedLanguage] = useState(null);
