+++ src/pages/SoilReport.jsx
@@ -1,13 +1,13 @@
-import yourLogo from "./assets/soil.png";
-import staffIcon from "./assets/staff.png";
-import desIcon from "./assets/des.png";
-import detectIcon from "./assets/detect.png";
-import uploadIcon from "./assets/upload.png";
-import Footer from "./components/Footer";
+import yourLogo from "../assets/soil.png";
+import staffIcon from "../assets/staff.png";
+import desIcon from "../assets/des.png";
+import detectIcon from "../assets/detect.png";
+import uploadIcon from "../assets/upload.png";
+import Footer from "../components/Footer";
 
 const SoilReport = () => {
   return (
-    <div className="h-screen flex justify-center flex-col bg-gray-200 items-center">
+    <div className="h-screen flex justify-center flex-col bg-gray-200 items-center w-full mx-auto">
       <div className="w-full h-full shadow-lg overflow-hidden">
         <div
           className="h-24"
@@ -17,14 +17,14 @@ const SoilReport = () => {
           }}
         ></div>
 
-        <div className="flex justify-center">
+        <div className="flex justify-center items max-w-md w-full mx-auto">
           <div
             className="h-40 bg-contain w-full max-w-xs bg-no-repeat rounded-lg"
             style={{ backgroundImage: `url(${yourLogo})` }}
           ></div>
         </div>
 
-        <div className="flex flex-col gap-4 p-4 rounded-lg">
+        <div className="flex flex-col gap-4 p-4 rounded-lg  max-w-md w-full mx-auto">
           <button className=" flex items-center  text-sm font-medium rounded-full  bg-white">
             <div className="bg-yellow-100 rounded-full p-2 flex items-center justify-center mr-2">
               <img src={staffIcon} alt="Report icon" className="h-8 w-8" />
