+++ src/pages/Step1.jsx
@@ -1,15 +1,16 @@
-import StepPage from "./StepPage";
-import uploadIcon from './assets/upload.png';
-import LocationIcon from './assets/location.png';
+import React from "react";
+import savpalImg from "../assets/savpal.png";
+import uploadIcon from "../assets/upload.png";
+import LocationIcon from "../assets/location.png";
+import StepPage from "../components/StepPage";
 
-const Step5 = () =>{
-
-    const pageTitle = "Step 5 : Apply weedicides as recommended" 
-    const pageInstruction = "Apply Urea based on recommended dose to support vegetative growth, once crop is healthy andweed-free."
-    return (
+const Step1 = () => {
+  const pageTitle = "Step 1 : Apply 500 kg FOM before field preparation";
+  const pageInstruction = "Apply 500 kg of Fermented Organic Manure (FOM) evenly across the land before field preparation. This improves soil organic carbon and supports microbial activity."
+   return (
         <StepPage title={pageTitle} instruction={pageInstruction}>
          
-               <div className="flex flex-col gap-4">
+               <div className="flex flex-col gap-2 w-full mx-auto">
              
                  <div className="flex items-center gap-2 bg-white rounded-full p-2 shadow-md">
          
@@ -18,7 +19,7 @@ const Step5 = () =>{
                      <img src={uploadIcon} alt="Seed photo icon" className="h-6 w-6" />
                    </div>
                    <span className="font-normal text-black text-sm flex-1">
-                     Urea photo
+                     Field photo
                    </span>
          
                    <label>
@@ -30,8 +31,8 @@ const Step5 = () =>{
                  </div>
          
                  {/* Current Location Input */}
-                 <div className="flex items-center gap-2 bg-white rounded-full p-2 shadow-md">
-                   <div className="bg-purple-200 p-2 rounded-full flex items-center justify-center">
+                <div className="flex items-center gap-2 bg-white rounded-full p-2 shadow-md">
+          <div className="bg-purple-200 p-2 rounded-full flex items-center justify-center">
                      <img
                        src={LocationIcon}
                        alt="Location icon"
@@ -48,7 +49,7 @@ const Step5 = () =>{
                </div>
          
                {/* Note Section */}
-               <div className="bg-white p-3 rounded-xl shadow-md">
+               <div className="bg-white p-2 rounded-lg shadow-md">
                  <label
                    htmlFor="notes"
                    className="block text-sm font-normal text-gray-700"
@@ -64,8 +65,8 @@ const Step5 = () =>{
                </div>
          
                {/* Submit Button */}
-               <button className="bg-[#2f6003] text-white font-semibold py-4 rounded-full shadow-lg w-full">
-                 Resubmit
+               <button className="bg-[#2f6003] text-white font-semibold py-2 rounded-full shadow-lg w-full">
+                 Submit
                </button>
 
 
@@ -76,17 +77,17 @@ const Step5 = () =>{
             Your query has been submitted
           </span>
         </p>
-        <p className="text-lg font-bold text-cyan-900 mt-2">
+         <p className="text-lg font-bold text-cyan-900 mt-2">
           Waiting Time: 00:00
         </p>
         <p className="text-gray-500 text-xs italic text-center mt-2">
-          Only proceed to the next step once your current step is verified and
-          approved by the staff.
+         You have  completed all the steps to get
+           update from the support
         </p>
       </div>
          
         </StepPage>
     )
-}
+};
 
-export default Step5;
+export default Step1;
